+++ widgets/device.js
@@ -67,6 +67,48 @@ $.widget("sv.device_codepad", $.sv.widget, {
 
 });
 
+
+// ----- device.roofwindow ------------------------------------------------------
+$.widget("sv.device_roofwindow", $.sv.widget, {
+
+	initSelector: 'div[data-widget="device.roofwindow"]',
+
+	options: {
+		min: 0,
+		max: 255,
+		step: 5,
+	},
+
+	_getVal: function(event) {
+		var min = this.options.min;
+		var max = this.options.max;
+		var step = this.options.step;
+
+		var offset = $(event.currentTarget).offset();
+		var y = event.pageY - offset.top;
+		return max - Math.floor(y / $(event.currentTarget).outerHeight() * (max - min) / step) * step;
+	},
+
+	_events: {
+		'click .pos': function (event) {
+			this._write(this._getVal(event));
+		},
+
+		'mouseenter .pos': function (event) {
+			this.element.find('.control').fadeIn(400);
+		},
+
+		'mouseleave .pos': function (event) {
+			this.element.find('.control').fadeOut(400);
+		},
+
+		'mousemove .pos': function (event) {
+			$(event.currentTarget).attr('title', this._getVal(event));
+		}
+	}
+
+});
+
   // ----- device.uzsu ----------------------------------------------------------
   // ----------------------------------------------------------------------------
   //