+++ changelog.md
@@ -10,6 +10,7 @@
 - New: basic.color (supersedes basic.colordisc and basic.rgb) with new parameter 'colormodel' for HSV or HSL model and possibility to pass values as list in one item
 - New: status.badge (displays a notification badge)
 - New: basic.offset (button to increase or decrease a value)
+- New: device.roofwindow (to show and control a roof window)
 - New: device.uzsuicon (to control UZSU in smarthome.py and FHEM)
 - New: device.uzsugraph (to control UZSU in smarthome.py and FHEM)
 - New: calendar.waste (compact view of waste collection dates)
@@ -19,6 +20,7 @@
 - New: icon.cistern
 - New: icon.garagedoor
 - New: icon.heating (displays a heating colored with dynamic gradient)
+- New: îcon.roofwindow
 - basic.symbol: Can also be used to show text only and to render as link, mode extended to adaptable formula (including thresholds), and - most important - may have multiple states now (so eventually, no series of symbols is needed anymore to cover mutiple states)
 - plot.period: Among other things: merged functionality of plot.minmaxavg and plot.multiaxis into it, more options like logarithmic and boolean scale, units, an advanced zoom mode as in Highstock, individual count and mode per series and the possibility to set any additional chart options
 - plot.temprose: New parameters 'series_label' and 'unit'